[#1922] Updated NextKey Response Value for Event Pagination
- Updated logic to set the next key if provided from service layer
- Added additional logging
