+++ cla-backend-go/v2/events/converters.go
@@ -1,0 +1,37 @@
+// Copyright The Linux Foundation and each contributor to CommunityBridge.
+// SPDX-License-Identifier: MIT
+
+package events
+
+import (
+	v1Models "github.com/communitybridge/easycla/cla-backend-go/gen/models"
+	"github.com/communitybridge/easycla/cla-backend-go/gen/v2/models"
+	"github.com/jinzhu/copier"
+)
+
+func v2EventList(eventList *v1Models.EventList) (*models.EventList, error) {
+	var dst models.EventList
+	err := copier.Copy(&dst, eventList)
+	if err != nil {
+		return nil, err
+	}
+	return &dst, nil
+}
+
+type codedResponse interface {
+	Code() string
+}
+
+func errorResponse(err error) *models.ErrorResponse {
+	code := ""
+	if e, ok := err.(codedResponse); ok {
+		code = e.Code()
+	}
+
+	e := models.ErrorResponse{
+		Code:    code,
+		Message: err.Error(),
+	}
+
+	return &e
+}
